+++ Public/HP/Get-RaidControllerHP.ps1
@@ -201,6 +201,41 @@ function Get-RaidControllerHP{
         $RAIDStatus = "Healthy"
     }
 
+    # Define regular expressions to extract information
+    $arrayRegex = 'Array\s+(\w+)'
+    $logicalDriveRegex = 'logicaldrive\s+(\d+)\s*\(([^,]+),\s*([^,]+),\s*([^,]+)(?:,\s*([^)]+))?\)'
+
+    # Define a function to extract information from a logical drive string
+    function Get-LogicalDriveInfo {
+        param([string]$array, [string]$logicalDriveString)
+        
+        if ($logicalDriveString -match $logicalDriveRegex) {
+            [PSCustomObject]@{
+                Array               = $array.Trim()
+                Name                = "logicaldrive $($matches[1])"
+                Size                = $matches[2].Trim()
+                RaidType            = $matches[3].Trim()
+                Status              = $matches[4].Trim()
+                Progress            = if ($matches[5]) { $matches[5].Trim() } else { $null }
+                RowColour           = if($matches[4].Trim() -ne "OK" ) {'warning'} else {'success'}
+            }
+        }
+    }
+
+    # Create a list to store logical drive information
+    $virtualdrives = New-Object System.Collections.Generic.List[Object]
+
+    # Extract logical drive information
+    $currentArray = $null
+    foreach ($line in $hpraidstatusslot_ld -split "`n") {
+        if ($line -match $arrayRegex) {
+            $currentArray = $matches[1]
+        }
+        elseif ($line -match 'logicaldrive') {
+            $virtualdrives.Add((Get-LogicalDriveInfo -array $currentArray -logicalDriveString $line))
+        }
+    }
+
     # Extract Percentage if rebuilding
     $rebuildpercentage = [regex]::Match($hpraidstatusslot_ld, '\d+\.\d+%').Value
 
@@ -213,5 +248,5 @@ function Get-RaidControllerHP{
         RowColour               = if (($RAIDStatus -eq 'Not Healthy') -or ($RAIDphysicalstatus -eq 'Not Healthy')) {"danger"}elseif ($rebuildpercentage -ne "") {'warning'}else{"success"}
     })
 
-    return $raidarraydetails, $AllDrives, $faileddrives
+    return $raidarraydetails, $AllDrives, $virtualdrives, $faileddrives
 }